The various GUIs are organized in patterns
