+++ src/main/java/edivad/dimstorage/client/screen/element/button/CollectButton.java
@@ -8,7 +8,7 @@ public class CollectButton extends Button {
 
 	public CollectButton(int width, int height, boolean isCollecting, Button.IPressable onPress)
 	{
-		super(width / 2 + 95, height / 2 + 70, 64, 20, getText(isCollecting), onPress);
+		super(width, height, 64, 20, getText(isCollecting), onPress);
 	}
 
 	private static String getText(boolean isCollecting)
