+++ src/main/java/edivad/dimstorage/client/screen/element/button/OwnerButton.java
@@ -7,7 +7,7 @@ public class OwnerButton extends Button {
 
 	public OwnerButton(int width, int height, String owner, Button.IPressable onPress)
 	{
-		super(width / 2 + 95, height / 2 - 57, 64, 20, owner, onPress);
+		super(width, height, 64, 20, owner, onPress);
 		this.active = Config.DIMCHEST_ALLOWPRIVATENETWORK.get();
 	}
 }
