+++ src/main/java/edivad/dimstorage/client/screen/element/textfield/FrequencyText.java
@@ -7,7 +7,7 @@ public class FrequencyText extends TextFieldWidget {
 
 	public FrequencyText(int width, int height, String msg)
 	{
-		super(Minecraft.getInstance().fontRenderer, width / 2 + 95, height / 2 - 15, 64, 15, msg);
+		super(Minecraft.getInstance().fontRenderer, width, height, 64, 15, msg);
 		setMaxStringLength(3);
 		setVisible(true);
 		setFocused(false);
