+++ src/main/java/edivad/dimstorage/client/screen/pattern/BaseScreen.java
@@ -1,7 +1,8 @@
-package edivad.dimstorage.client.screen;
+package edivad.dimstorage.client.screen.pattern;
 
 import com.mojang.blaze3d.platform.GlStateManager;
 
+import edivad.dimstorage.tools.Translate;
 import net.minecraft.client.gui.screen.inventory.ContainerScreen;
 import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.inventory.container.Container;
@@ -10,12 +11,14 @@ import net.minecraft.util.text.ITextComponent;
 
 public class BaseScreen<T extends Container> extends ContainerScreen<T> {
 
-	protected ResourceLocation background;
+	private ResourceLocation background;
 
 	public BaseScreen(T container, PlayerInventory invPlayer, ITextComponent text, ResourceLocation background)
 	{
 		super(container, invPlayer, text);
 		this.background = background;
+		this.xSize = 176;
+		this.ySize = 220;
 	}
 
 	@Override
@@ -23,6 +26,13 @@ public class BaseScreen<T extends Container> extends ContainerScreen<T> {
 	{
 		GlStateManager.color4f(1.0F, 1.0F, 1.0F, 1.0F);
 		this.minecraft.getTextureManager().bindTexture(background);
+	}
+
+	@Override
+	protected void drawGuiContainerForegroundLayer(int mouseX, int mouseY)
+	{
+		this.font.drawString(this.getTitle().getFormattedText(), 8, 6, 4210752);
+		this.font.drawString(Translate.translateToLocal("container.inventory"), 8, 128, 4210752);
 	}
 
 	@Override
