+++ src/main/java/edivad/dimstorage/tile/TileFrequencyOwner.java
@@ -18,9 +18,12 @@ import net.minecraftforge.api.distmarker.OnlyIn;
 
 public abstract class TileFrequencyOwner extends TileEntity implements ITickableTileEntity, INamedContainerProvider {
 
+	public boolean locked;
+	
 	public TileFrequencyOwner(TileEntityType<?> tileEntityTypeIn)
 	{
 		super(tileEntityTypeIn);
+		locked = false;
 	}
 
 	public Frequency frequency = new Frequency();
@@ -29,7 +32,7 @@ public abstract class TileFrequencyOwner extends TileEntity implements ITickable
 	public void setFreq(Frequency frequency)
 	{
 		this.frequency = frequency;
-		markDirty();
+		this.markDirty();
 		BlockState state = world.getBlockState(pos);
 		world.notifyBlockUpdate(pos, state, state, 3);
 	}
@@ -41,6 +44,12 @@ public abstract class TileFrequencyOwner extends TileEntity implements ITickable
 			setFreq(frequency.copy().setPublic());
 		else
 			setFreq(frequency.copy().setOwner(Main.proxy.getClientPlayer()));
+	}
+	
+	public void swapLocked()
+	{
+		locked = !locked;
+		this.markDirty();
 	}
 
 	public boolean canAccess(PlayerEntity player)
@@ -65,6 +74,7 @@ public abstract class TileFrequencyOwner extends TileEntity implements ITickable
 	{
 		super.read(tag);
 		frequency.set(new Frequency(tag.getCompound("Frequency")));
+		locked = tag.getBoolean("locked");
 	}
 
 	@Override
@@ -72,6 +82,7 @@ public abstract class TileFrequencyOwner extends TileEntity implements ITickable
 	{
 		super.write(tag);
 		tag.put("Frequency", frequency.writeToNBT(new CompoundNBT()));
+		tag.putBoolean("locked", locked);
 		return tag;
 	}
 
@@ -90,6 +101,7 @@ public abstract class TileFrequencyOwner extends TileEntity implements ITickable
 	{
 		CompoundNBT tag = super.getUpdateTag();
 		tag.put("Frequency", frequency.writeToNBT(new CompoundNBT()));
+		tag.putBoolean("locked", locked);
 		return tag;
 	}
 
@@ -97,5 +109,6 @@ public abstract class TileFrequencyOwner extends TileEntity implements ITickable
 	public void handleUpdateTag(CompoundNBT tag)
 	{
 		frequency.set(new Frequency(tag.getCompound("Frequency")));
+		locked = tag.getBoolean("locked");
 	}
 }
